Atualizando a estrutura

+++ VilaoEletrico/VilaoEletrico/ObjetoCasa.cs
@@ -5,6 +5,6 @@
         public string Nome { get; set; }
         public string Local { get; set; }
         public double ConsumoWatts { get; set; }
-        public int TempoUsoDia { get; set; }
+        public double TempoUsoDia { get; set; }
     }
 }
